some refactoring, added tabs to the apis page.
