+++ projects/ui/src/Components/ApiDetails/SchemaTab/redoc/RedocDisplay.tsx
@@ -1,5 +1,5 @@
 import { RedocStandalone } from "redoc";
-import { APISchema } from "../../Apis/api-types";
+import { APISchema } from "../../../../Apis/api-types";
 import { RedocDisplayContainer } from "./RedocDisplay.style";
 
 export function RedocDisplay({ spec }: { spec: APISchema | undefined }) {
