+++ projects/ui/src/Components/ApiDetails/SchemaTab/swagger/SwaggerDisplay.style.tsx
@@ -1,8 +1,8 @@
 import { css } from "@emotion/react";
 import styled from "@emotion/styled";
-import { contentWidth } from "../../Styles/ContentWidthHelpers";
-import { borderRadiusConstants } from "../../Styles/constants";
-import { makeStyledButtonCSS } from "../Common/Button";
+import { contentWidth } from "../../../../Styles/ContentWidthHelpers";
+import { borderRadiusConstants } from "../../../../Styles/constants";
+import { makeStyledButtonCSS } from "../../../Common/Button";
 
 /*
 *   EXPLANATION OF div[id^="display-swagger"]
