+++ projects/ui/src/Components/Apis/PendingSubscriptionsTab/PendingSubscriptionTabContent.tsx
@@ -1,0 +1,5 @@
+const PendingSubscriptionTabContent = () => {
+  return <div>PendingSubscriptionTabContent</div>;
+};
+
+export default PendingSubscriptionTabContent;
