+++ projects/ui/tsconfig.json
@@ -22,7 +22,8 @@
     "./src/**/*.svg",
     "./src/**/*.ts",
     "./src/**/*.tsx",
-    ".eslintrc.js"
+    ".eslintrc.js",
+    ".tsconfig.json"
   ],
   "exclude": ["node_modules"],
   "references": [{ "path": "./tsconfig.node.json" }]
